pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "feedparser>=6.0.11",
    "flask-socketio>=5.5.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "cryptography>=44.0.2",
    "scikit-learn>=1.6.1",
    "python-jwt>=4.1.0",
    "psutil>=7.0.0",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "nltk>=3.9.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.3",
]
